+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Use strlcpy to save session replay info path ()
+
 ## 8.44.0-beta.1
 
 ### Fixes